Adding WPF Example to ReadMe
Adding a Drop In WPF Sample to the ReadMe. 

(Although this may seem unnecessary for experienced C#/WPF, those new to C#/WPF may appreciate a drop in working sample code with the library. I personally find this one of the most helpful aspects coming from the Ruby community when utilizing gems in a project -- Saving 30 minutes in poking around/testing can be the difference in whether I choose to use I'll use the library.)